+++ R/tbl-df.r
@@ -97,10 +97,10 @@ as.data.frame.tbl_df <- function(x, row.names = NULL, optional = FALSE, ...) {
 
 #' @rdname dplyr-formatting
 #' @export
-print.tbl_df <- function(x, ..., n = NULL) {
+print.tbl_df <- function(x, ..., n = NULL, width = NULL) {
   cat("Source: local data frame ", dim_desc(x), "\n", sep = "")
   cat("\n")
-  trunc_mat(x, n = n)
+  trunc_mat(x, n = n, width = width)
 }
 
 #' @export