+++ R/tbl-dt.r
@@ -59,10 +59,10 @@ as.data.frame.tbl_dt <- function(x, row.names = NULL, optional = FALSE, ...) {
 
 #' @export
 #' @rdname dplyr-formatting
-print.tbl_dt <- function(x, ..., n = NULL) {
+print.tbl_dt <- function(x, ..., n = NULL, width = NULL) {
   cat("Source: local data table ", dim_desc(x), "\n", sep = "")
   cat("\n")
-  trunc_mat(x, n = n)
+  trunc_mat(x, n = n, width = width)
 }
 
 #' @export